visualization scripts moved to beam_to_matsim/visualization/via/
