+++ src/main/scala/beam/utils/beam_to_matsim/visualization/via/visualization_20_24.scala
@@ -1,8 +1,8 @@
-package beam.utils.beam_to_matsim.visualization
+package beam.utils.beam_to_matsim.visualization.via
 
 import beam.utils.beam_to_matsim.events.BeamPathTraversal
-import beam.utils.beam_to_matsim.io.{Reader, Writer}
 import beam.utils.beam_to_matsim.events_filter.{MutablePopulationFilter, MutableSamplingFilter, PopulationSample}
+import beam.utils.beam_to_matsim.io.{Reader, Writer}
 import beam.utils.beam_to_matsim.via_event.ViaEvent
 
 object visualization_20_24 extends App {
