+++ src/main/scala/beam/utils/beam_to_matsim/visualization/via/visualization_9.scala
@@ -1,8 +1,8 @@
-package beam.utils.beam_to_matsim.visualization
+package beam.utils.beam_to_matsim.visualization.via
 
 import beam.utils.beam_to_matsim.events.BeamPathTraversal
-import beam.utils.beam_to_matsim.io.{HashSetReader, Reader, Writer}
 import beam.utils.beam_to_matsim.events_filter.{MutablePopulationFilter, MutableSamplingFilter, PopulationSample}
+import beam.utils.beam_to_matsim.io.{HashSetReader, Reader, Writer}
 import beam.utils.beam_to_matsim.via_event.ViaEvent
 
 import scala.collection.mutable
